types: move final data into own module
